% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{binary.mc}
\alias{binary.mc}
\title{Pricing Barrier option using Monte Carlo method}
\usage{
binary.mc(obj, env, n, steps, all = FALSE, plot = FALSE, ...)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{steps}{A number specifying the number of steps each asset price trajectory will contain, used only for \code{method = "mc"}.}

\item{all}{A logical value specifying whether the pricing function should return only the result price (\code{all = FALSE}) or other data during computation (\code{all = TRUE}). The default value for this argument is \code{FALSE}.}

\item{plot}{A logical value specifying whether the simulated price trajectories should be plotted. The default value for this argument is \code{FALSE}.}

\item{...}{Graphical arguments used for plotting.}
}
\description{
Pricing Barrier option using Monte Carlo method
}
\keyword{internal}
