% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{Binomial.Basic}
\alias{Binomial.Basic}
\title{Basic n-step Binomial Options Pricing Model}
\usage{
Binomial.Basic(
  K,
  S,
  u,
  d,
  r,
  t,
  n,
  type = "call",
  style = "European",
  all = FALSE,
  plot = FALSE
)
}
\arguments{
\item{K}{A number representing the exercise price of the option.}

\item{S}{A number representing the current price of the stock.}

\item{u}{A number representing the ratio of rise in price in the binomial tree.}

\item{d}{A number representing the ratio of fall in price in the binomial tree.}

\item{r}{A number representing the risk-free annual interest rate.}

\item{t}{A number representing the number of years reaching maturity.}

\item{n}{An integer representing the number of steps of the binomial model.}

\item{type}{A string specifying the type of options, takes values "call" or "put, default takes value "call".}

\item{style}{A string specifying the style of options, takes values "European" or "American", default takes value "European".}

\item{plot}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether the binomial tree should be plotted.}
}
\value{
A number representing the computed current price of the option.
}
\description{
Basic n-step Binomial Options Pricing Model
}
\examples{
Binomial.Basic(K = 14, S = 8, u = 2, d = 0.5, r = 0.25, t = 0.75, n = 3)
}
