% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{Binomial}
\alias{Binomial}
\title{Cox, Ross & Rubinstein Lattice Binomial Options Pricing Model}
\usage{
Binomial(
  K,
  S,
  u = exp(sigma * sqrt(t/n)),
  d = exp(-sigma * sqrt(t/n)),
  r,
  t,
  n,
  sigma = 0,
  type = "call",
  style = "European",
  all = FALSE,
  plot = FALSE
)
}
\arguments{
\item{K}{A number representing the exercise price of option.}

\item{S}{A number representing the current price of option, this is also the asset value after a horizontal jump.}

\item{u}{A number representing the ratio of an up jump in the asset value.}

\item{d}{A number representing the ratio of an down jump in the asset value.}

\item{r}{A number representing the continuously compounded yearly interest rate.}

\item{t}{A number representing the time to option maturity (in years).}

\item{n}{An integer representing the number of time steps into which the interval of length t is divided.}

\item{sigma}{A number representing the volatility (in standard deviation) of the rate of return on the underlying asset (yearly).}

\item{type}{A character evaluating to either \code{"call"} or \code{"put"}, representing the type of the option; the default value is set to be \code{"call"}.}

\item{style}{A character evaluating to either \code{"European"} or \code{"American"}, representing the style of the option; the default value is set to be \code{"American"}.}

\item{all}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether the output should contain all option price, risk-neutral probability, and stock price tree, or option price only; the default value is set to be \code{FALSE}.}

\item{plot}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether the binomial tree should be plotted; the default value is set to be \code{FALSE}.}
}
\value{
If \code{all = FALSE}, return only a number representing the result option price, where as \code{all = TRUE} return a list containing the result options price, risk-neutral probability, and the stock price tree (matrix). If \code{plot = TRUE}, the function will also produce a plot of the Binomial tree.
}
\description{
Cox, Ross & Rubinstein Lattice Binomial Options Pricing Model
}
\examples{
Binomial(14, 8, 2, 0.5, 0.25, 0.75, 3, all = TRUE, type = "put", style = "American")
}
