% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{Binomial.rnv}
\alias{Binomial.rnv}
\title{Backward Recursive Risk-neutral Valuation for Binomial Model}
\usage{
Binomial.rnv(
  fn,
  p,
  r,
  n,
  dt,
  K = NULL,
  S = NULL,
  type = "call",
  style = "European"
)
}
\arguments{
\item{fn}{A numeric vector containing all possible stock prices at the final step/state of the binomial tree.}

\item{p}{A number representing the risk-neutral probability.}

\item{r}{A number representing the continuously compounded yearly interest rate.}

\item{dt}{A number representing the length in time of each step.}

\item{K}{A number representing the exercise price of option, needed only when calculating an American option.}

\item{S}{A number representing the current price of option, needed only when calculating an American option.}

\item{type}{A character evaluating to either \code{"call"} or \code{"put"}, representing the type of the option; the default value is set to be \code{"call"}.}

\item{style}{A character evaluating to either \code{"European"} or \code{"American"}, representing the style of the option; the default value is set to be \code{"American"}.}
}
\value{
A number representing the result options price.
}
\description{
Backward Recursive Risk-neutral Valuation for Binomial Model
}
\keyword{internal}
