% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{Binomial.tree}
\alias{Binomial.tree}
\title{Generating Binomial Tree Matrix}
\usage{
Binomial.tree(S0, u, d, n)
}
\arguments{
\item{S0}{A number representing the current price of the underlying asset.}

\item{u}{A number representing the ratio of an upward price movement.}

\item{d}{A number representing the ratio of an downward price movement.}

\item{n}{An integer representing the number of time steps into which the interval of length t is divided.}
}
\value{
A matrix representing the Binomial lattice tree.
}
\description{
Generating Binomial Tree Matrix
}
\examples{
rmcop:::Binomial.tree(S0 = 8, u = 2, d = 0.5, n = 3)
}
\keyword{internal}
