% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackScholes.R
\name{BlackScholes.d1}
\alias{BlackScholes.d1}
\title{Compute d1 in Black-Scholes Formula}
\usage{
BlackScholes.d1(S, K, t, r, q, sigma)
}
\arguments{
\item{S}{A number representing the current price of the underlying asset.}

\item{K}{A number representing the strike price of the option.}

\item{t}{A number representing the time to maturity in years.}

\item{r}{A number representing the continuously compounded risk-free interest rate per annum.}

\item{q}{A number representing the continuously compounded dividend yield per annum.}

\item{sigma}{A number representing the volatility per annum.}
}
\value{
A number representing the value of d1 used in Black-Scholes formula.
}
\description{
Compute d1 in Black-Scholes Formula
}
\keyword{internal}
