% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackScholes.R
\name{BlackScholes}
\alias{BlackScholes}
\title{Black-Scholes Model for Option Pricing}
\usage{
BlackScholes(S, K, t, r, q, sigma, type = "call")
}
\arguments{
\item{S}{A number representing the current price of the underlying asset.}

\item{K}{A number representing the strike price of the option.}

\item{t}{A number representing the time to maturity in years.}

\item{r}{A number representing the continuously compounded risk-free interest rate per annum.}

\item{q}{A number representing the continuously compounded dividend yield per annum.}

\item{sigma}{A number representing the volatility per annum.}

\item{type}{A character evaluating to either \code{"call"} or \code{"put"}, representing the type of the option; the default value is set to be \code{"call"}.}
}
\value{
A number representing the current price of the option.
}
\description{
Black-Scholes Model for Option Pricing
}
\keyword{internal}
