% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{check.args}
\alias{check.args}
\title{Object integrity check for price.option functions}
\usage{
check.args(obj, env, option)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}
}
\description{
Object integrity check for price.option functions
}
\keyword{internal}
