% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{check.method}
\alias{check.method}
\title{Check method supporting status}
\usage{
check.method(method)
}
\arguments{
\item{method}{The specified method to check if is supported by the package.}
}
\description{
Check method supporting status
}
\keyword{internal}
