% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{check.option}
\alias{check.option}
\title{Check option object integrity}
\usage{
check.option(obj)
}
\arguments{
\item{obj}{The defined option class object.}
}
\description{
Check option object integrity
}
\keyword{internal}
