% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{d1}
\alias{d1}
\title{Compute d2 in Black-Scholes Formulas}
\usage{
d1(t, sigma, S, K, r, q = 0)
}
\arguments{
\item{t}{A number representing time to expiration (in years).}

\item{sigma}{A number representing volatility per annual.}

\item{S}{A number representing the price of the underlying asset.}

\item{K}{A number representing the strike price.}

\item{r}{A number representing the continuously compounded risk-free interest rate per annual.}

\item{q}{A number representing the continuously compounded dividend yield per annual.}
}
\description{
Compute d2 in Black-Scholes Formulas
}
