% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{d2}
\alias{d2}
\title{Compute d2 in Black-Scholes Formulas}
\usage{
d2(t, sigma, d1 = NA, S = NULL, K = NULL, r = NULL, q = 0)
}
\arguments{
\item{t}{A number representing time to expiration (in years).}

\item{sigma}{A number representing volatility per annual.}

\item{d1}{A number representing the d1 value, if this argument is specified, it will directly used to calculate the value of d2, and the following arguments will not be used.}

\item{S}{A number representing the price of the underlying asset.}

\item{K}{A number representing the strike price.}

\item{r}{A number representing the continuously compounded risk-free interest rate per annual.}

\item{q}{A number representing the continuously compounded dividend yield per annual.}
}
\description{
Compute d2 in Black-Scholes Formulas
}
