% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.R
\name{fv}
\alias{fv}
\title{Future Value of A Plain Vanilla Bond}
\usage{
fv(pv, r, m, n = 1)
}
\arguments{
\item{pv}{a number representing the present value of an asset at a specific point in time in the future.}

\item{r}{a number representing the fixed annual interest rate.}

\item{m}{an optional number specifying the number of interest payments within a year.}

\item{n}{a number representing the number of years forward.}
}
\value{
If \code{m} is not specified, the future value is computed via continuous compounding using natural logarithm.

If \code{m} is specified as a number, the future value is computed via discrete compounding.
}
\description{
\code{fv} computes the future value of a plain vanilla bond given its present value and assumes a fixed interest rate.
}
\examples{
fv(pv = 100, r = 0.05, m = 1) # = 105
}
