% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{mc.engine}
\alias{mc.engine}
\title{Engine for Monte Carlo Simulation on Asset Price Trajectory}
\usage{
mc.engine(type, K, t, S0, r, q, sigma, n, steps)
}
\arguments{
\item{type}{A string specifying the type of the option of interest. Possible values are either \code{"call"} or \code{"put"}, corresponding to call option (right to buy) and pull option (right to sale), respectively.}

\item{K}{A number specifying the strike price of the option.}

\item{t}{A number specifying the maturity/expiry of the option in years.}

\item{S0}{A number specifying the current price of the underlying asset.}

\item{r}{A number specifying the (fixed) annual interest rate.}

\item{q}{A number specifying the (fixed) annual dividend yield rate of the option.}

\item{sigma}{A number specifying the annual volatility measure.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{steps}{A number specifying the number of steps each asset price trajectory will contain, used only for \code{method = "mc"}.}
}
\value{
A matrix recording the simulated price trajectories.
}
\description{
Engine for Monte Carlo Simulation on Asset Price Trajectory
}
\keyword{internal}
