% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{mc.SE}
\alias{mc.SE}
\title{Compute Monte Carlo Model's Standard Error}
\usage{
mc.SE(C_i, C.mean, n)
}
\arguments{
\item{C_i}{A matrix recording the resulted option pay-offs from Monte Carlo simulations.}

\item{C.mean}{A number recording the resulted estimate of the price of the option.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}
}
\value{
A number recording the standard error of the corresponding Monte Carlo estimate of the option price.
}
\description{
Compute Monte Carlo Model's Standard Error
}
\keyword{internal}
