% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.env}
\alias{option.env}
\title{Initialise new \code{env} class object}
\usage{
option.env(method = "mc", S, r, q = 0, sigma, n = NULL, steps = NULL)
}
\arguments{
\item{method}{A string specifying the method used to price the option. Supported values are \code{"mc"} (Monte Carlo Method), \code{"bs"} (for Black-Scholes formula), \code{"binomial"} (Binomial Lattice Tree).}

\item{S}{A number specifying the current price of the underlying asset.}

\item{r}{A number specifying the (fixed) annual interest rate.}

\item{q}{A number specifying the (fixed) annual dividend yield rate of the option.}

\item{sigma}{A number specifying the annual volatility measure.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{steps}{A number specifying the number of steps each asset price trajectory will contain, used only for \code{method = "mc"}.}
}
\value{
An S3 object with \code{class} attribute evaluated to \code{"env"}. It represents the market environment or the properties of the underlying asset for the option pricing setting.
}
\description{
Create an S3 class object which characterised the market environment used when pricing the option of interest. The returned object is used together with an \code{"option"} class object as inputs of the pricing function \code{price.option}.
}
\examples{
# Create an environment with current price 20, annual interest rate 2\%, annual
# dividend yield rate 1\%, annual volatility measure 0.05, and number of simulations
# set to 100, with each trajectory breaking into 100 steps.
option.env(S = 20, r = 0.02, q = 0.01, sigma = 0.05, n = 100, steps = 100)

# Alternatively, one can leave arguments `n` and `steps` unspecified or as NULL,
# and specify the two parameters when calling the `price.option` function.
option.env(S = 20, r = 0.02, q = 0.01, sigma = 0.05)
}
