% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.exotic.asian}
\alias{option.exotic.asian}
\title{Initialise Asian option}
\usage{
option.exotic.asian(obj, is.avg_price = TRUE)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{is.avg_price}{A logical value for \verb{"asian" "option"} class objects, specifying whether the option is an "average price Asian option" or "average strike Asian option". The default value for this argument is \code{TRUE}, indicating the option being an "average price Asian option".}
}
\description{
Initialise Asian option
}
\keyword{internal}
