% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.exotic.barrier}
\alias{option.exotic.barrier}
\title{Initialise barrier option}
\usage{
option.exotic.barrier(obj, barrier, is.knockout = TRUE)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{barrier}{A number for \verb{"barrier" "option"} class objects, specifying the barrier price of the barrier option.}

\item{is.knockout}{A logical value for \verb{"barrier" "option"} class objects, specifying whether the option is an "knock-out barrier option" (\code{is.knockout = TRUE}) or "knock-in barrier option" (\code{is.knockout = FALSE}).}
}
\description{
Initialise barrier option
}
\keyword{internal}
