% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.exotic.binary}
\alias{option.exotic.binary}
\title{Initialise binary option}
\usage{
option.exotic.binary(obj, payout)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{payout}{A number for \verb{"binary" "option"} class objects, specifying the payout of a binary option.}
}
\description{
Initialise binary option
}
\keyword{internal}
