% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.exotic.lookback}
\alias{option.exotic.lookback}
\title{Initialise lookback option}
\usage{
option.exotic.lookback(obj, is.fixed = TRUE)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{is.fixed}{A logical value for \verb{"lookback" "option"} class objects. specifying whether the option is an "fixed strike look-back option" (\code{is.fixed = TRUE}) or "floating strike look-back option" (\code{is.fixed = FALSE}).}
}
\description{
Initialise lookback option
}
\keyword{internal}
