% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.exotic}
\alias{option.exotic}
\title{Initialise exotic option}
\usage{
option.exotic(obj, option, ...)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{option}{A string specifying the specific family of the option of interest. Supported values are: \code{"vanilla"} (default, for Vanilla options), \code{"asian"} (for Asian options), \code{"barrier"} (for Barrier options), \code{"binary"} (for Binary/Cash-or-nothing options), and \code{"lookback"} (for Lookback options).}

\item{...}{Arguments required by corresponding specified option families.}
}
\description{
Initialise exotic option
}
\keyword{internal}
