% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option}
\alias{option}
\title{Initialise new \code{option} class object}
\usage{
option(style, option = "vanilla", type, K, t, ...)
}
\arguments{
\item{style}{A string specifying the style of the option of interest. Possible values are \code{"european"} and \code{"american"}, corresponding to European options and American options, respectively.}

\item{option}{A string specifying the specific family of the option of interest. Supported values are: \code{"vanilla"} (default, for Vanilla options), \code{"asian"} (for Asian options), \code{"barrier"} (for Barrier options), \code{"binary"} (for Binary/Cash-or-nothing options), and \code{"lookback"} (for Lookback options).}

\item{type}{A string specifying the type of the option of interest. Possible values are either \code{"call"} or \code{"put"}, corresponding to call option (right to buy) and pull option (right to sale), respectively.}

\item{K}{A number specifying the strike price of the option.}

\item{t}{A number specifying the maturity/expiry of the option in years.}

\item{...}{Arguments required by corresponding specified option families.}
}
\value{
An S3 object with \code{class} attribute evaluated to \code{"option"}.
}
\description{
Initialise new \code{option} class object
}
\examples{
# Create an European vanilla option with strike price 20, expiry in 1 year.
vanilla1 <- option("european", "vanilla", "call", K = 20, t = 1)

# Create European  average strike Asian option with strike price 20, expiry in 1 year.
asian1 <- option("european", "asian", "call", K = 20, t = 1, is.avg_price = FALSE)

# Create European knock-out barrier option with strike price 20, barrier price 21, expiry in 1 year.
barrier1 <- option("european", "barrier", "call", K = 20, t = 1, barrier = 21, is.knockout = TRUE)

# Create European binary option with strike price 20, fixed payout 5, expiry in 1 year.
binary1 <- option("european", "binary", "call", K = 20, t = 1, payout = 5)

# Create European fixed strike lookback option with strike price 20, expiry in 1 year.
lookback1 <- option("european", "lookback", "call", K = 20, t = 1, is.fixed = TRUE)
}
