% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option.vanilla}
\alias{option.vanilla}
\title{Initialise vanilla option}
\usage{
option.vanilla(obj)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}
}
\description{
Initialise vanilla option
}
\keyword{internal}
