% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price.R
\name{price.option.bs}
\alias{price.option.bs}
\title{Option Pricing via Black-Scholes Model}
\usage{
price.option.bs(obj, env, n, all)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{all}{A logical value specifying whether the pricing function should return only the result price (\code{all = FALSE}) or other data during computation (\code{all = TRUE}). The default value for this argument is \code{FALSE}.}
}
\description{
Option Pricing via Black-Scholes Model
}
\keyword{internal}
