% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Title Present Value of A Bond}
\usage{
pv(fv, r, m, n = 1)
}
\arguments{
\item{fv}{a number representing the future value of an asset at a specific point in time in the future.}

\item{r}{a number representing the fixed annual interest rate.}

\item{m}{an optional number specifying the number of interest payments within a year.}

\item{n}{a number representing the number of years reaching the specified future value \code{fv}.}
}
\value{
If \code{m} is not specified, the present value is computed via continuous compounding using natural logarithm.

If \code{m} is specified as a number, the present value is computed via discrete compounding.
}
\description{
\code{pv} computes the present value of a plain vanilla bond given its future value and assumes a fixed interest rate.
}
\examples{
pv(fv = 105, r = 0.05, m = 1) # = 1
}
