% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{rnv.american}
\alias{rnv.american}
\title{Risk-neutral Valuation for American Options with Recursion Method}
\usage{
rnv.american(fn, p, r, dt, K, S, type = "call")
}
\arguments{
\item{fn}{A vector containing the final states value of the Binomial model.}

\item{p}{A number representing the risk-neutral probability for the stock price to rise.}

\item{r}{A number representing the fixed annual interest rate.}

\item{dt}{A number representing the time lapse between each Binomial sub-tree.}

\item{K}{A number representing the strike price}

\item{S}{A matrix representing the stock price tree.}

\item{type}{A string specifying the type of option, either "call" or "put", with "put" being the default option.}
}
\value{
A number representing the computed current value of the option.
}
\description{
Risk-neutral Valuation for American Options with Recursion Method
}
\keyword{internal}
