% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{rnv.european}
\alias{rnv.european}
\title{Risk-neutral Valuation for European Options}
\usage{
rnv.european(fn, p, r, dt)
}
\arguments{
\item{fn}{A vector containing the final states value of the Binomial model.}

\item{p}{A number representing the risk-neutral probability for the stock price to rise.}

\item{r}{A number representing the fixed annual interest rate.}

\item{dt}{A number representing the time lapse between each Binomial sub-tree.}
}
\value{
A number representing the computed current value of the option.
}
\description{
Risk-neutral Valuation for European Options
}
\keyword{internal}
