% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{seq_geom}
\alias{seq_geom}
\title{Geometric Sequence for Vectors using Recursion Method}
\usage{
seq_geom(x, q, n)
}
\arguments{
\item{x}{An vector used as the base of calculation.}

\item{q}{An fixed ratio.}

\item{n}{A number specifying the number of times we wish to advanced for the geometric sequence.}
}
\value{
A vector concatenating the original vector and the resulting geometric sequences.
}
\description{
Geometric Sequence for Vectors using Recursion Method
}
\keyword{internal}
