% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trinomial.R
\name{Trinomial.rnv}
\alias{Trinomial.rnv}
\title{Backward Recursive Risk-neutral Valuation for Trinomial Model}
\usage{
Trinomial.rnv(
  fn,
  p1,
  p2,
  p3,
  r,
  n,
  dt,
  K = NULL,
  S = NULL,
  type = "call",
  style = "european"
)
}
\arguments{
\item{fn}{A numeric vector containing all possible stock prices at the final step/state of the binomial tree.}

\item{p1}{A number representing the risk-neutral probability of an upward movement of the asset price.}

\item{p2}{A number representing the risk-neutral probability of an horizontal movement of the asset price.}

\item{p3}{A number representing the risk-neutral probability of an downward movement of the asset price.}

\item{r}{A number representing the (assumed fixed) continuously compounded yearly interest rate.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{dt}{A number representing the length in time of each step.}

\item{K}{A number representing the exercise price of option, needed only when calculating an american option.}

\item{S}{A number representing the current price of option, needed only when calculating an american option.}

\item{type}{A character evaluating to either \code{"call"} or \code{"put"}, representing the type of the option; the default value is set to be \code{"call"}.}

\item{style}{A character evaluating to either \code{"european"} or \code{"american"}, representing the style of the option; the default value is set to be \code{"american"}.}
}
\value{
A number representing the result options price.
}
\description{
Backward Recursive Risk-neutral Valuation for Trinomial Model
}
\keyword{internal}
