% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{update.env}
\alias{update.env}
\title{Updating \code{option.env} class object}
\usage{
\method{update}{env}(
  env,
  method = env$method,
  S = env$S,
  r = env$r,
  q = env$q,
  sigma = env$sigma,
  n = env$n,
  steps = env$steps
)
}
\arguments{
\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}

\item{method}{A string specifying the method used to price the option. Supported values are \code{"mc"} (Monte Carlo Method), \code{"bs"} (for Black-Scholes formula), \code{"binomial"} (Binomial Lattice Tree), and \code{"trinomial"} (Trinomial Lattice Tree).}

\item{S}{A number specifying the current price of the underlying asset.}

\item{r}{A number specifying the (fixed) annual interest rate.}

\item{q}{A number specifying the (fixed) annual dividend yield rate of the option.}

\item{sigma}{A number specifying the annual volatility measure.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{steps}{A number specifying the number of steps each asset price trajectory will contain, used only for \code{method = "mc"}.}
}
\value{
An S3 object with \code{class} attribute evaluated to \code{"env"}.
}
\description{
Updating \code{option.env} class object
}
\examples{
# Create an environment with current price 20, annual interest rate 2\%,
# annual dividend yield rate 1\%, annual volatility measure 0.05, and number
# of simulations set to 100
env1 <- option.env(S = 20, r = 0.02, q = 0.01, sigma = 0.05, n = 100)

# Update env1's current price S to 25
env1 <- update.env(env1, S = 25)
}
