% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{vanilla.binomial}
\alias{vanilla.binomial}
\title{Binomial model for pricing vanilla options}
\usage{
vanilla.binomial(obj, env, n, u, d, all = FALSE)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{u}{A number representing the ratio of an upward price movement.}

\item{d}{A number representing the ratio of an downward price movement.}

\item{all}{A logical value specifying whether the pricing function should return only the result price (\code{all = FALSE}) or other data during computation (\code{all = TRUE}). The default value for this argument is \code{FALSE}.}
}
\description{
Binomial model for pricing vanilla options
}
\keyword{internal}
