% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackScholes.R
\name{vanilla.bs}
\alias{vanilla.bs}
\title{Pricing vanilla option using Black-Scholes model}
\usage{
vanilla.bs(obj, env, all = FALSE)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}

\item{all}{A logical value specifying whether the pricing function should return only the result price (\code{all = FALSE}) or other data during computation (\code{all = TRUE}). The default value for this argument is \code{FALSE}.}
}
\description{
Pricing vanilla option using Black-Scholes model
}
\keyword{internal}
