% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{check.args}
\alias{check.args}
\title{Check argument inputs integrity for price.option function}
\usage{
check.args(obj, env, option)
}
\arguments{
\item{obj}{The specified \code{"option"} class object which encapsulate some properties of an option of interest.}

\item{env}{The specified \code{"env"} class object which encapsulate some market variables required by corresponding pricing methods.}
}
\description{
Check argument inputs integrity for price.option function
}
\keyword{internal}
