% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{mc.engine}
\alias{mc.engine}
\title{Engine for Monte Carlo Simulation on Asset Price Trajectory}
\usage{
mc.engine(t, S0, r, q, sigma, n, steps)
}
\arguments{
\item{t}{A number specifying the maturity/expiry of the option in years.}

\item{S0}{A number specifying the current price of the underlying asset.}

\item{r}{A number specifying the (fixed) annual interest rate.}

\item{q}{A number specifying the (fixed) annual dividend yield rate of the option.}

\item{sigma}{A number specifying the annual volatility measure.}

\item{n}{A number speciying the number of simulations to make (for \code{method = "mc"}), or the number of time steps the life of the option will be broken into (for \code{method = "binomial"} and \code{method = "trinomial"}).}

\item{steps}{A number specifying the number of steps each asset price trajectory will contain, used only for \code{method = "mc"}.}
}
\value{
A matrix recording the simulated price trajectories. Noticing that if dividend rate q is not zero, the resulted asset price trajectories are ex-dividend. When a stock goes ex-dividend, its price typically drops by the amount of the dividend. This is because once the stock goes ex-dividend, new buyers of the stock are no longer entitled to receive the dividend payment.
}
\description{
Engine for Monte Carlo Simulation on Asset Price Trajectory
}
\examples{
# Example simulating price trajectories, noticing that if dividend rate q is not zero,
# the resulted asset price trajectories are ex-dividend.
with_dividend <- rmcop:::mc.engine(t = 2, S0 = 20, r = 0.5, q = 0.1, sigma = 0.1,
n = 100, steps = 30)
no_dividend <- rmcop:::mc.engine(t = 2, S0 = 20, r = 0.5, q = 0, sigma = 0.1,
n = 100, steps = 30)

# Plot trajectories
rmcop:::mc.plot(with_dividend)
rmcop:::mc.plot(no_dividend)
}
\keyword{internal}
