% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{mc.plot}
\alias{mc.plot}
\title{Monte Carlo Price Trajectory Plot}
\usage{
mc.plot(S, threshold, ...)
}
\arguments{
\item{...}{Other parameters used for plotting}

\item{S_i}{Price trajectory matrix.}
}
\description{
Monte Carlo Price Trajectory Plot
}
\keyword{internal}
