% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{option}
\alias{option}
\title{Initialise new \code{option} class object}
\usage{
option(style, option = "vanilla", type, K, t, ...)
}
\arguments{
\item{style}{A string specifying the style of the option of interest. Possible values are \code{"european"} and \code{"american"}, corresponding to European options and American options, respectively.}

\item{option}{A string specifying the specific family of the option of interest. Supported values are: \code{"vanilla"} (default, for Vanilla options), \code{"asian"} (for Asian options), \code{"barrier"} (for Barrier options), \code{"binary"} (for Binary/Cash-or-nothing options), and \code{"lookback"} (for Lookback options). This argument is a string that specifies the subclass of the S3 class \code{"option"} for the defined object.}

\item{type}{A string specifying the type of the option of interest. Possible values are either \code{"call"} or \code{"put"}, corresponding to call option (right to buy) and pull option (right to sale), respectively.}

\item{K}{A number specifying the strike price of the option.}

\item{t}{A number specifying the maturity/expiry of the option in years.}

\item{...}{Arguments required by corresponding specified option families.
\itemize{
\item \code{is.avg_price}: A logical value for \code{option = "asian"} case, specifying whether the option is an "average price Asian option" or "average strike Asian option". The default value for this argument is \code{TRUE}, indicating the option being an "average price Asian option".
\item \code{barrier}: A number for \verb{option = "barrier} case, specifying the barrier price of the barrier option.
\item \code{is.knockout}: A logical value for \code{option = "barrier"} class objects, specifying whether the option is an "knock-out barrier option" (\code{is.knockout = TRUE}) or "knock-in barrier option" (\code{is.knockout = FALSE}).
\item \code{payout}: A number for \code{option = "binary"} class objects, specifying the payout of a binary option.
\item \code{is.fixed}: A logical value for \code{option = "lookback"} class objects. specifying whether the option is an "fixed strike look-back option" (\code{is.fixed = TRUE}) or "floating strike look-back option" (\code{is.fixed = FALSE}).
}}
}
\value{
An S3 object with \code{class} attribute evaluated to \code{"option"}. This object can then be used for pricing together with an \code{"env"} class object through \code{price.option()} function.
}
\description{
Create an S3 class object which characterised an option of interest. The returned object is then used together with a \code{"env"} class object together as inputs of the \code{price.option} function for further option pricing.
}
\examples{
# Create an European vanilla option with strike price 20, expiry in 1 year.
option("european", "vanilla", "call", K = 20, t = 1)

# Create European average strike Asian option with strike price 20, expiry in 1 year.
option("european", "asian", "call", K = 20, t = 1, is.avg_price = FALSE)

# Create European knock-out barrier option with strike price 20, barrier price 21, expiry in 1 year.
option("european", "barrier", "call", K = 20, t = 1, barrier = 21, is.knockout = TRUE)

# Create European binary option with strike price 20, fixed payout 5, expiry in 1 year.
option("european", "binary", "call", K = 20, t = 1, payout = 5)

# Create European fixed strike lookback option with strike price 20, expiry in 1 year.
option("european", "lookback", "call", K = 20, t = 1, is.fixed = TRUE)
}
